requirements.txt
streamlit
pymupdf
openai
python-dotenv